awesome itoa_base -> pointer OK

+++ ft_pointer.c
@@ -6,7 +6,7 @@
 /*   By: hthomas <hthomas@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/29 17:49:15 by hthomas           #+#    #+#             */
-/*   Updated: 2019/12/04 16:33:42 by hthomas          ###   ########.fr       */
+/*   Updated: 2019/12/04 20:26:28 by hthomas          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	ft_pointer(va_list arg, t_sp *sp)
 {
-	sp->p = va_arg(arg, unsigned long long);
+	sp->p = va_arg(arg, long);
 	sp->len += ft_count_len_hex(sp->p) + 2;
 	return ('p');
 }
